apps/api/pyproject.toml
[project]
name = "esui-api"
version = "0.0.1"
description = "ESUI backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pgvector>=0.3.0",
    "python-socketio>=5.11.0",
    "redis>=5.2.0",
    "anthropic>=0.40.0",
    "voyageai>=0.3.0",
    "httpx>=0.27.0",
    "python-jose[cryptography]>=3.3.0",
    "email-validator>=2.2.0",
    "python-multipart>=0.0.12",
    "boto3>=1.35.0",
    "apscheduler>=3.10.0",
    "structlog>=24.4.0",
    "resend>=2.4.0",
    "pillow>=10.4.0",
    "feedparser>=6.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.7.0",
    "mypy>=1.13.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
